// src/app/api/fn/retell-calls/retell-calls-controller-find-all.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { CallListResponse } from '../../models/call-list-response';

export interface RetellCallsControllerFindAll$Params {
  search?: string;
  callStatus?: string;
  userSentiment?: 'Negative' | 'Positive' | 'Neutral' | 'Unknown';
  page?: number;
  limit?: number;
  status?: 'active' | 'suspended' | 'all';
  withTrashed?: (boolean | 'true' | 'false');
  onlyTrashed?: (boolean | 'true' | 'false');
  start_date?: string;
  end_date?: string;
}

export function retellCallsControllerFindAll(http: HttpClient, rootUrl: string, params?: RetellCallsControllerFindAll$Params, context?: HttpContext): Observable<StrictHttpResponse<CallListResponse>> {
  const rb = new RequestBuilder(rootUrl, retellCallsControllerFindAll.PATH, 'get');
  if (params) {
    rb.query('search', params.search, {});
    rb.query('callStatus', params.callStatus, {});
    rb.query('userSentiment', params.userSentiment, {});
    rb.query('page', params.page, {});
    rb.query('limit', params.limit, {});
    rb.query('status', params.status, {});
    rb.query('withTrashed', params.withTrashed, {});
    rb.query('onlyTrashed', params.onlyTrashed, {});
    rb.query('start_date', params.start_date, {});
    rb.query('end_date', params.end_date, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<CallListResponse>;
    })
  );
}

retellCallsControllerFindAll.PATH = '/api/v1/retell/calls';